// src/api/models/codeInfoOutput.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * EdCraft Backend API
 * API for EdCraft Backend
 * OpenAPI spec version: 0.1.0
 */
import type { BranchElement } from './branchElement';
import type { CodeTreeOutput } from './codeTreeOutput';
import type { FunctionElement } from './functionElement';
import type { LoopElement } from './loopElement';

/**
 * Model representing code structure and code elements.
 */
export interface CodeInfoOutput {
  /** A hierarchical representation of the code structure. */
  code_tree: CodeTreeOutput;
  /** List of available functions in the code. */
  functions: FunctionElement[];
  /** List of available loops in the code. */
  loops: LoopElement[];
  /** List of available branches in the code. */
  branches: BranchElement[];
  /** List of variable names used in the code. */
  variables: string[];
}