// src/api/models/codeTreeOutput.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * EdCraft Backend API
 * API for EdCraft Backend
 * OpenAPI spec version: 0.1.0
 */

/**
 * Model representing a node in the code tree.
 */
export interface CodeTreeOutput {
  /** Unique identifier for the code element. */
  id: number;
  /** Type of the code element (function, loop, branch). */
  type: string;
  /** List of variable names in the current code element. */
  variables: string[];
  /** List of indices of functions within this code element. */
  function_indices: number[];
  /** List of indices of loops within this code element. */
  loop_indices: number[];
  /** List of indices of branches within this code element. */
  branch_indices: number[];
  /** List of child code elements. */
  children: CodeTreeOutput[];
}